+++ peloton-frontend/component/home/homemain/RaceItem.js
