+++ peloton-frontend/component/home/homemain/RaceItemImage.js
