+++ peloton-frontend/component/home/homemain/RaceItemText.js
