+++ peloton-frontend/component/home/homemain/RaceItems.js
