+++ peloton-frontend/component/home/homemain/RaceListTitle.js
